+++ src/PolyFeature.cpp
@@ -37,19 +37,15 @@
 #endif
 
 PolyFeature::PolyFeature()
-        : TopoFeature(), _poly(), _groundElevations(), _polyInternalID(),
+        : TopoFeature(), _poly(), _groundElevations(), _polyInternalID(new_internal_id()),
       _groundElevation(-global::largnum), _minBbox() {}
 
 PolyFeature::PolyFeature(const int outputLayerID)
-    : TopoFeature(outputLayerID), _poly(), _groundElevations(), _polyInternalID(),
+        : TopoFeature(outputLayerID), _poly(), _groundElevations(), _polyInternalID(new_internal_id()),
       _groundElevation(-global::largnum), _minBbox() {}
 
-PolyFeature::PolyFeature(const int outputLayerID, const int internalID)
-        : TopoFeature(outputLayerID), _groundElevations(), _polyInternalID(internalID),
-          _groundElevation(-global::largnum), _minBbox() {}
-
 PolyFeature::PolyFeature(const nlohmann::json& poly, const bool checkSimplicity)
-        : TopoFeature(), _groundElevations(), _polyInternalID(),
+        : TopoFeature(), _groundElevations(), _polyInternalID(new_internal_id()),
       _groundElevation(-global::largnum), _minBbox() {
     this->parse_json_poly(poly, checkSimplicity);
 }
@@ -63,19 +59,8 @@ PolyFeature::PolyFeature(const nlohmann::json& poly, const bool checkSimplicity,
 PolyFeature::PolyFeature(const nlohmann::json& poly, const int outputLayerID)
         : PolyFeature(poly, false, outputLayerID) {}
 
-PolyFeature::PolyFeature(const nlohmann::json& poly, const bool checkSimplicity,
-                         const int outputLayerID, const int internalID)
-        : PolyFeature(poly, checkSimplicity) {
-    _polyInternalID = internalID;
-    _outputLayerID  = outputLayerID;
-    if (_outputLayerID  >= _numOfOutputLayers) _numOfOutputLayers = _outputLayerID + 1;
-}
-
-PolyFeature::PolyFeature(const nlohmann::json& poly, const int outputLayerID, const int internalID)
-        : PolyFeature(poly, false, outputLayerID, internalID) {}
-
 PolyFeature::PolyFeature(const Polygon_with_attr& poly, const bool checkSimplicity)
-        : TopoFeature(), _groundElevations(), _polyInternalID(),
+        : TopoFeature(), _groundElevations(), _polyInternalID(new_internal_id()),
           _groundElevation(-global::largnum), _minBbox() {
     bool isOuterRing = true;
     for (auto& ring : poly.polygon.rings()) {
@@ -122,18 +107,9 @@ PolyFeature::PolyFeature(const Polygon_with_attr& poly, const bool checkSimplici
 PolyFeature::PolyFeature(const Polygon_with_attr& poly, const int outputLayerID)
         : PolyFeature(poly, false, outputLayerID) {}
 
-PolyFeature::PolyFeature(const Polygon_with_attr& poly, const bool checkSimplicity,
-                         const int outputLayerID, const int internalID)
-        : PolyFeature(poly, checkSimplicity) {
-    _polyInternalID = internalID;
-    _outputLayerID  = outputLayerID;
-    if (_outputLayerID  >= _numOfOutputLayers) _numOfOutputLayers = _outputLayerID + 1;
-}
-
-PolyFeature::PolyFeature(const Polygon_with_attr& poly, const int outputLayerID, const int internalID)
-        : PolyFeature(poly, false, outputLayerID, internalID) {}
-
 PolyFeature::~PolyFeature() = default;
+
+int PolyFeature::_numOfPolyFeatures = 0;
 
 void PolyFeature::calc_footprint_elevation_nni(const DT& dt) {
     typedef std::vector<std::pair<DT::Point, double>> Point_coordinate_vector;
@@ -201,8 +177,8 @@ void PolyFeature::calc_footprint_elevation_linear(const DT& dt) {
 
 double PolyFeature::ground_elevation() {
     if (_groundElevation < -global::largnum + global::smallnum) {
-        if (_groundElevations.empty())throw std::runtime_error("Polygon elevations missing!"
-                                                           " Cannot calculate average");
+        if (_groundElevations.empty()) throw std::runtime_error("Polygon elevations missing!"
+                                                                " Cannot calculate average");
         // calculating base elevation as 95 percentile of outer ring
         _groundElevation = geomutils::percentile(_groundElevations.front(), 0.95);
     }
@@ -301,12 +277,22 @@ void PolyFeature::clear_feature() {
     _mesh.clear();
 }
 
+int PolyFeature::new_internal_id() {
+    return ++_numOfPolyFeatures;
+}
+
 Polygon_with_holes_2& PolyFeature::get_poly() {
     return _poly;
 }
 
 const Polygon_with_holes_2& PolyFeature::get_poly() const {
     return _poly;
+}
+
+Polygon_with_attr PolyFeature::get_poly_w_attr() const {
+    Polygon_with_attr poly;
+    poly.polygon = _poly;
+    return poly;
 }
 
 const std::vector<std::vector<double>>& PolyFeature::get_ground_elevations() const {