+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [Unreleased]
 ### Added
 - Precompiled Windows executable
+### Changed
+- Improved fallbacks in case building import fails in late stages
+### Fixed
+- Minor bugfixes
 
 ## [0.4.3] - 2023-08-25
 ### Fixed